Addressing some Sonar blocker issues

+++ src/v2i-hub/CswPlugin/src/VehicleLocate.cpp
@@ -74,7 +74,7 @@ int VehicleLocate::FindRegion(TiDataFrame *frame, WGS84Point point, uint16_t hea
 		int nodesLength = geoPath->description->choice.path.offset.choice.xy.choice.nodes.list.count;
 
 		WGS84Point* points = GetPointArray(nodes, nodesLength, nodesAnchor);
-		bool inList = IsInPointList(points, nodesLength, point, heading, laneWidth/100.0);
+		bool inList = IsInPointList(points, nodesLength, point, heading, double(laneWidth)/100.0);
 		free(points);
 		if (inList) {
 			return i + 1;
